% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Daily weather data from the US Sheep Experiment Station}
\format{A data frame with 101783 rows and 3 variables:
\describe{
  \item{date}{date in \%Y-\%m-\%d format}
  \item{ELEMENT}{weather variable: PRCP, TMAX, TMIN}
  \item{value}{value of weather variable, PRCP 0.1 mm, TMAX (0.1 C), TMIN (0.1 C)}
}}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/all/USC00102707.dly}
}
\usage{
weather
}
\description{
A dataset containing the date, the daily max, min and mean
mean temperatures, and the total rainfall and snowfall for
that date. Data recorded at the US Sheep Experiment Station
weather station GHCND:USC00102707
}
\keyword{datasets}
